feat: 2.0 rewrite
BREAKING CHANGE: the `tabbableElements` option can now be a selector, a list of elements, or a function that returns a list of elements.

BREAKING CHANGE: removed support for jQuery

BREAKING CHANGE: removed static `trapAll`, replacing it with static `create`, which simply creates an instance. To replicate the functionality of `trapAll`, a `trapFocus` export has been added.